* commit '499e1de2509c0afd8daaa135c3ffdd29ab2b74df': (46 commits)
  #9324 Fixed: mandatory setting on on multple choice answer does not work mandatory setting on quiz does not allow to go back to previous question and then forward again. Long text answers options overlaps each other
  risolto #4576 - Link a numeri in Forma LMS
  risolto #8990 - Link scorretto in Gestione amministratori - Annulla
  #4229 riportata anche gestione customfield per corsi e edizioni - cambiato layout
  Revert #8791 Call sl_uploads method in lib.multimedia.php
  #1547: fixed mandatory answer in case of more questions in a single page
  #1547: Fixed. The red alert does not disappear when user select the answer
  bug fix #9171   , in caso di domande-test pescate da un sottinsieme di categorie
  Risolto errore: non salva testo customfield di tipo text se metto testo (salvava solo numeri)
  Modificato titolo pagina in amministrazione customfield
  #9163: Fixed uncorrect use of php preg_replace function
  #4229 riportata anche gestione customfield per corsi e edizioni
  #4229 Risolto problema cerca non funzionante in question bank
  Risolti errori di ordinamento in question bank
  fixed: 9093 Risolto problema a preg_replace come indicato in segnalazione.
  #4229 riportata anche gestione customfield per corsi e edizioni, sistemata gestione customfield per campo di tipo testo
  Risolto #4229 Inserita la gestione campi custom nei test e nel raccoglitore delle domande
  removed useless variable assignment
  fixed: #9048 Era un problema causato dalla modifica al menu dropdown che avevo fatto.
  fixed: #9002 Corretto bottone modifica nascosto
  ...

# Conflicts:
#	html/appLms/modules/question/class.question.php
#	html/appLms/modules/question/class.text_entry.php
#	html/composer.lock

+++ html/appLms/admin/models/CourseAlms.php
@@ -692,6 +692,13 @@ Class CourseAlms extends Model
         if ($error)
             $res['err'] = '_err_course';
         else {
+            
+            // Salvataggio CustomField
+            require_once(_adm_.'/lib/lib.customfield.php');
+            $extra_field = new CustomFieldList();
+            $extra_field->setFieldArea( "COURSE" );
+            $extra_field->storeFieldsForObj( $id_course );
+    
             //AUTO SUBSCRIPTION
             if (isset($_POST['auto_subscription']) && $_POST['auto_subscription'] == 1) {
                 $userId = Docebo::user()->getIdSt();
@@ -1006,6 +1013,12 @@ Class CourseAlms extends Model
 		if($quota_exceeded)
 			$res['limit_reach'] = 1;
         
+        // Salvataggio CustomField
+        require_once(_adm_.'/lib/lib.customfield.php');
+        $extra_field = new CustomFieldList();
+        $extra_field->setFieldArea( "COURSE" );
+        $extra_field->storeFieldsForObj( $id_course );
+    
         //AUTO SUBSCRIPTION
         $userId = Docebo::user()->getIdSt();
         $userSubscribed = $this->isUserSubscribedInCourse($userId, $id_course);
@@ -1270,7 +1283,9 @@ Class CourseAlms extends Model
 		$cmodel->deleteAllCourseCompetences($id_course);
 		//--- end competences ------------------------------------------------------
 
-
+                //remove customfield
+		if(!sql_query("DELETE FROM ".$GLOBALS['prefix_fw']."_customfield_entry WHERE id_field IN (SELECT id_field FROM core_customfield WHERE area_code = 'COURSE') AND id_obj = '".$id_course."'"))
+			return false;
 
 		//--- finally delete course from courses table -----------------------------
 		if(!sql_query("DELETE FROM %lms_course WHERE idCourse = '".$id_course."'"))
